Create IsArraySorted.java
Take as input N, the size of array. Take N more inputs and store that in an array. Write a recursive function which returns true if the array is sorted and false otherwise. Print the value returned.

Input Format
Enter a number N and take N more inputs and store in an array

Constraints
None

Output Format
Display the Boolean result

Sample Input
4
1
2
3
4
Sample Output
true